Separate speaker assignment from diarization
