+++ scinoephile/audio/__init__.py
@@ -5,6 +5,6 @@
 Package hierarchy (modules may import from any above):
 * cache_namespace / samples
 * separation / vad
-* transcription
-* diarization / subtitles
+* diarization / transcription
+* speaker_assignment / subtitles
 """
